requirements.txt
# Meta Aria 2 Surgical Training Analysis
# Requirements file

# Core dependencies
numpy>=1.24.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Project Aria Tools (for Meta Aria 2 glasses)
projectaria-tools

# Visualization
rerun-sdk

# Optional: Machine Learning for custom detection
# Uncomment if you want to train custom surgical tool detection models
# torch>=2.0.0
# torchvision>=0.15.0
# transformers>=4.30.0

# Optional: Additional analysis tools
# seaborn>=0.12.0
# plotly>=5.14.0